+++ src/defaults/default.config.js
@@ -4,6 +4,7 @@ module.exports = {
 	templateDir: 'templates',
 	staticDir: 'static',
 	distDir: 'dist',
+	filterDir: 'filters',
 	base: '/',
 	templateExt: 'html',
 	permalinks: {