+++ src/filters/format_date.js
@@ -1,0 +1,5 @@
+const dayjs = require('dayjs');
+
+module.exports = function(date, format) {
+	return dayjs(date).format(format);
+};