+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 -   Adds support for `.js` data files ([#42](https://github.com/marceljs/marcel/issues/42)).
 -   Adds `build` (default) and `serve` commands to the CLI ([#22](https://github.com/marceljs/marcel/issues/22)).
+-   Adds support for sync/async custom filters ([#43](https://github.com/marceljs/marcel/issues/43)). For this, we switched to async rendering for Nunjucks.
+-   Adds the `format_date` default filter.
 
 ### 0.1.3